# api/docs/swagger.yaml
definitions:
...
  itineraries.Activities:
    properties:
      text:
        type: string
    type: object
  itineraries.Author:
    properties:
      full_name:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  itineraries.CommentItinerariesReq:
    properties:
      author_id:
        type: string
      content:
        type: string
      itinerary_id:
        type: string
    type: object
  itineraries.CommentItinerariesRes:
    properties:
      author_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      itinerary_id:
        type: string
    type: object
  itineraries.Destination:
    properties:
      activities:
        items:
          $ref: '#/definitions/itineraries.Activities'
        type: array
      end_date:
        type: string
      name:
        type: string
      start_date:
        type: string
    type: object
  itineraries.GetItinerariesByIdRes:
    properties:
      author:
        $ref: '#/definitions/itineraries.Author'
      description:
        type: string
      destination:
        items:
          $ref: '#/definitions/itineraries.Destination'
        type: array
      end_date:
        type: string
      id:
        type: string
      start_date:
        type: string
      title:
        type: string
    type: object
  itineraries.GetItinerariesRes:
    properties:
      itineraries:
        items:
          $ref: '#/definitions/itineraries.ItinerariesRes'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  itineraries.ItinerariesReq:
    properties:
      description:
        type: string
      destinations:
        items:
          $ref: '#/definitions/itineraries.Destination'
        type: array
      end_date:
        type: string
      start_date:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
  itineraries.ItinerariesRes:
    properties:
      created_at:
        type: string
      description:
        type: string
      end_date:
        type: string
      id:
        type: string
      start_date:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
  itineraries.UpdateItinerariesReq:
    properties:
      description:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/itineraries:
    get:
      description: Get itineraries
      parameters:
      - description: Number of itineraries to fetch
        in: query
        name: limit
        type: string
      - description: Number of itineraries to omit
        in: query
        name: offset
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/itineraries.GetItinerariesRes'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get
      tags:
      - itineraries
    post:
      description: create itineraries
      parameters:
      - description: info
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/itineraries.ItinerariesReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/itineraries.ItinerariesRes'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: create
      tags:
      - itineraries
  /api/v1/itineraries/{itinerary_id}:
    delete:
      description: Delete itineraries
      parameters:
      - description: itinerary_id
        in: path
        name: itinerary_id
        required: true
        type: string
      responses:
        "200":
          description: successfully deleted
          schema:
            type: string
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete
      tags:
      - itineraries
    get:
      description: Get itineraries by id
      parameters:
      - description: itinerary_id
        in: path
        name: itinerary_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/itineraries.GetItinerariesByIdRes'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get
      tags:
      - itineraries
    put:
      description: update itineraries
      parameters:
      - description: itinerary_id
        in: path
        name: itinerary_id
        required: true
        type: string
      - description: info
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/itineraries.UpdateItinerariesReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/itineraries.ItinerariesRes'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: update
      tags:
      - itineraries
  /api/v1/itineraries/{itinerary_id}/comments:
    post:
      description: comment itineraries
      parameters:
      - description: itinerary_id
        in: path
        name: itinerary_id
        required: true
        type: string
      - description: info
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/itineraries.CommentItinerariesReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/itineraries.CommentItinerariesRes'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: comment
      tags:
      - itineraries
...
securityDefinitions:
  ApiKeyAuth:
    description: API Gateway
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"